+++ libraries/WaterMix/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2023-08-30
+- add **void addExact(float volume, float temperature)** (works only for water).
+- add **float mass()** (works only for water).
+- add **float volume2mass(float volume, float temperature)** (works only for water).
+- add **float mass2volume(float mass, float temperature)** (works only for water).
+- update readme.md
+- fix keywords.txt
+- minor edits
+
+
 ## [0.1.0] - 2023-08-29
 - initial version